requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
PyNaCl==1.5.0